moved root file to src
