+++ test/backboneTypes.js
@@ -1,4 +1,4 @@
-    var Nested = require( '../nestedtypes' ),
+    var Nested = require( '../src/nestedtypes' ),
         expect = require( 'chai' ).expect,
         _ = require( 'underscore' ),
         sinon = require( 'sinon' );
